+++ Runtime/ProtoCode/HitCollider.cs
@@ -11,13 +11,13 @@ namespace ProtoCode
         public EffectDefinition effectDefinition => m_EffectDefinition;
 
         // 원래는 플레이어가 생성하면서 초기화 함수로 본인을 넣어줘야함
-        [SerializeField] IEntity m_Actor;
-        public IEntity actor => m_Actor;
+        [SerializeField] IStatController m_Actor;
+        public IStatController actor => m_Actor;
 
         /// <summary>
         /// 생성될 때, 스텟을 받도록 할수도 있음
         /// </summary>
-        public void Initialize(IEntity actorEntity)
+        public void Initialize(IStatController actorEntity)
         {
             m_Actor = actorEntity;
         }
@@ -32,10 +32,11 @@ namespace ProtoCode
 
         public void Invoke<T>(T target) where T : Component, IStatReceiver
         {
-            IEntity entity = target as IEntity;
+            IStatController entity = target as IStatController;
             var stat = entity.statHandler.GetHealthStat();
 
-            m_Actor.statHandler.ApplyEffect(effectDefinition, stat);
+            //m_Actor.statHandler.ApplyInstantEffect(effectDefinition, stat);
+            m_Actor.statHandler.CreateEffectInstance(effectDefinition, stat);
             Debug.Log($"target Helth: {stat.currentValue}/{stat.finalValue}");
 
             if (stat.currentValue <= 0)