+++ Runtime/ProtoCode/ProjectileInvoker.cs
@@ -11,14 +11,14 @@ namespace ProtoCode
         public EffectDefinition effectDefinition => m_EffectDefinition;
 
         // 원래는 플레이어가 생성하면서 초기화 함수로 본인을 넣어줘야함
-        [SerializeField] IEntity m_Actor;
+        [SerializeField] IStatController m_Actor;
         [SerializeField] MonoBehaviour m_ActorMono;
-        public IEntity actor => m_Actor;
+        public IStatController actor => m_Actor;
 
         /// <summary>
         /// 생성될 때, 스텟을 받도록 할수도 있음
         /// </summary>
-        public void Initialize(IEntity actorEntity)
+        public void Initialize(IStatController actorEntity)
         {
             m_Actor = actorEntity;
         }
@@ -33,13 +33,11 @@ namespace ProtoCode
 
         public void Invoke<T>(T target) where T : Component, IStatReceiver
         {
-            IEntity entity = target as IEntity;
+            IStatController entity = target as IStatController;
             var stat = entity.statHandler.GetHealthStat();
-            Debug.Log($"Before Effect -> Health: {stat.currentValue}/{stat.finalValue}");
             
-            IEntity owner = m_ActorMono as IEntity;
-            owner.statHandler.ApplyEffect(effectDefinition, stat);
-            Debug.Log($"After Effect -> Health: {stat.currentValue}/{stat.finalValue}");
+            IStatController owner = m_ActorMono as IStatController;
+            owner.statHandler.CreateEffectInstance(effectDefinition, stat);
 
             Destroy(gameObject);
         }