+++ Runtime/ProtoCode/StatController.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 
 namespace ProtoCode
 {
-    public class StatController : MonoBehaviour, IEntity, IStatReceiver
+    public class StatController : MonoBehaviour, IStatController, IStatReceiver
     {
         [SerializeField] StatDatabase m_StatDatabase;
         public StatDatabase statDatabase => m_StatDatabase;
@@ -20,12 +20,11 @@ namespace ProtoCode
 
 
         // 디버깅 이후에 사리질 수 있음
-        [SerializeField] EffectDefinition testEffect;
+        //[SerializeField] EffectDefinition testEffect;
 
 
         void Awake()
         {
-            Debug.Log($"스텟 초기화 진행");
             InitializeStat();
         }
 
@@ -33,18 +32,16 @@ namespace ProtoCode
         {
             m_StatHandler = new StatHandler(m_StatDatabase);
             m_StatHandler.InitializeStat();
-
-            // 스텟 테스트
-            foreach (var stat in m_StatHandler.stats)
-            {
-                Debug.Log($"{stat.Key}");
-            }
-
         }
 
         public void Accept(IStatInvoker invoker)
         {
             invoker.Invoke(this);
+        }
+
+        public void CheckHealth()
+        {
+            
         }
     }
 }