+++ Runtime/Stats/Attribute.cs
@@ -1,4 +1,5 @@
 using System;
+using Dave6.StatSystem.Effect;
 
 namespace Dave6.StatSystem.Stat
 {
@@ -9,5 +10,6 @@ namespace Dave6.StatSystem.Stat
     public class Attribute : BaseStat
     {
         public Attribute(StatDefinition definition) : base(definition) { }
+
     }
 }