+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-## [0.0.7] - 2025.12.10
+## [0.0.9] - 2025.12.28
 
 ### Added
-- Effect에 OutputMultiplier 추가
+- EffectDefinition CustomEditor 추가
+
+### Changed
+- Modifier와 Effect 개념을 합침
+- Effect가 3가지 적용모드를 사용하도록 구조를 변경 (instant, periodic, sustained)
+
 
 ### Fixed
-- ResourceStat의 currentValue 클램프 버그 수정
+- ResourceStat의 currentValue 값을 지속 변화에 대응할 수 있게
+onCurrentValueChanged 이벤트 사용